Remove `polkadot_node_subsystem` module and move still used code into `collation_generation` module
