+++ polkadot/node/overseer/src/lib.rs
@@ -60,7 +60,6 @@
 #![allow(clippy::all)]
 
 pub mod collation_generation;
-mod polkadot_node_subsystem;
 
 use std::{
 	collections::{hash_map, HashMap},
